pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael",email = "michael@octave.nz"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.10,<4.0)",
    "django-filter (>=25.2,<26.0)",
    "pygments (>=2.19.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
